requirements.txt
requests
pandas
python-dotenv
duckdb
pyarrow
